Allow specifying whether a property is an identity for an entity
By expanding the `IKeyPropertyResolver` interface, it is now possible to indicate whether a key property is an identity for an entity. By default the resolved key property is an identity. Custom implementations of `IKeyPropertyResolver` can override this behavior. For example, Dapper.FluentMap.Dommel will provide an implementation where you can configure whether a property is an identity.

Resolves #49