Cache column and tables using full type name as cache key
Previously just the Name property of the type was used as cache key. This results in the same cache key for different nested classes with the same name.

Added some tests to verify the correct tables or columns are mapped from the Resolvers class.

Use the full name in the logs as well for clarity.