Allow inserts/updates of properties without a public setter (#199)
This behavior changed in 367ac90a to skip readonly properties in insert/update queries. However this imposes issues with POCO's containing properties with private setters which represent columns in the database and are updated via methods. Given this it seems reasonable to support these cases and do not allow for readonly properties which have no associated database column, methods should be used in that case. Added some test coverage for these cases.

Fixes #194