Fix the parameter prefix of insert queries
The BuildInsertQuery method was not using the ISqlBuilder.PrefixParameter method to add the prefix to the parameters.
It was instead always adding the "@" prefix.
Since the methods BuildUpdateQuery and BuildDeleteQuery where using the PrefixParameter, I made the same in the BuildInsertQuery method.
Also, in order to create a test case, I made some changes to the method BuildInsertQuery, to make it accessible in the ParameterPrefixTest class.