Remove caching in the custom type map provider
Previously the type map provider used a caching key which resulted in duplicate names for common structures such as:
- Type `Product` with a column `CategoryId` results in a
- Type `ProductCategory` with a column `Id`
Both the `CategoryId` and `Id` columns with their respective types resulted in a cache key of `ProductCategoryId` which in turn results in Dapper mapping data from the data reader to the wrong property on the type.

Since Dapper caches the generated type deserializes caching the property mappings is actually not necessary.